properly rename test files and arangobench test case files
