+++ client-tools/Benchmark/test-cases.h
@@ -23,14 +23,14 @@
 
 #pragma once
 
-#include "testcases/AqlInsertTest.h"
-#include "testcases/CustomQueryTest.h"
-#include "testcases/CollectionCreationTest.h"
-#include "testcases/DocumentCreationTest.h"
-#include "testcases/DocumentCrudAppendTest.h"
-#include "testcases/DocumentCrudTest.h"
-#include "testcases/DocumentCrudWriteReadTest.h"
-#include "testcases/DocumentImportTest.h"
-#include "testcases/EdgeCrudTest.h"
-#include "testcases/PersistentIndexTest.h"
-#include "testcases/VersionTest.h"
+#include "testcases/AqlInsertTestCase.h"
+#include "testcases/CustomQueryTestCase.h"
+#include "testcases/CollectionCreationTestCase.h"
+#include "testcases/DocumentCreationTestCase.h"
+#include "testcases/DocumentCrudAppendTestCase.h"
+#include "testcases/DocumentCrudTestCase.h"
+#include "testcases/DocumentCrudWriteReadTestCase.h"
+#include "testcases/DocumentImportTestCase.h"
+#include "testcases/EdgeCrudTestCase.h"
+#include "testcases/PersistentIndexTestCase.h"
+#include "testcases/VersionTestCase.h"
